[~]: Настройка панели администратора
