+++ backend/users/admin.py
@@ -4,7 +4,13 @@
 from django.contrib import admin
 from django.contrib.auth.admin import UserAdmin
 
-from users.models import CustomUser
+from users.models import Blog, CustomUser
+
+
+class BlogInline(admin.StackedInline):
+    model = Blog
+    can_delete = False
+    verbose_name_plural = 'Blog'
 
 
 @admin.register(CustomUser)
@@ -12,4 +18,4 @@ class CustomUserAdmin(UserAdmin):
     """
     Этот класс расширяет встроенный UserAdmin для дополнительной настройки в административном интерфейсе Django.
     """
-    pass
+    inlines = (BlogInline,)
